third-commit

+++ lib/models/shop_app/login_file.dart
@@ -17,7 +17,7 @@ class UserData {
   String? phone;
   String? image;
   int? points;
-  int? credit;
+  double? credit;
   String? token;
   UserData({
     this.id,
